+++ packages/workdo/PurchaseOrder/src/Http/Requests/UpdatePurchaseOrderRequest.php
@@ -1,0 +1,3 @@
+<?php
+namespace Workdo\PurchaseOrder\Http\Requests;
+class UpdatePurchaseOrderRequest extends StorePurchaseOrderRequest {}